[jsk_recognition_utils] Separate grid_plane.h from geo_util.h
